refactor(sdk): move hashes to utils folder
- Move hashes object from testing/helpers.ts to utils/hashes.ts
- Export hashes from utils/index.ts
- Update imports in pool.ts, discovery.ts, testing/index.ts
