+++ sdk/src/testing/helpers.ts
@@ -2,11 +2,7 @@
  * Shared helpers for testing utilities.
  */
 
-import type { CallAndProof, Proof, StarknetAddress } from "../interfaces.js";
-import { Witness } from "../interfaces.js";
-import type { Hash, PrivateKey, PublicKey } from "../utils/crypto.js";
-import { hash } from "../utils/crypto.js";
-import type { NoteNonce } from "../internal/index.js";
+import type { CallAndProof, Proof } from "../interfaces.js";
 
 // ============ Mock Helpers ============
 
@@ -33,32 +29,3 @@ export function createMockCallAndProof(overrides?: Partial<CallAndProof>): CallA
 
 // Symbol used as a type marker for withdrawal operations (vs NoteNonce for transfers)
 export const Withdrawal = Symbol("Withdrawal");
-
-// ============ Hash Helpers ============
-
-export const hashes = {
-  channelKey: (
-    from: StarknetAddress,
-    fromPrivateKey: PrivateKey,
-    to: StarknetAddress,
-    toPublicKey: PublicKey
-  ): Hash => hash(from, fromPrivateKey, to, toPublicKey),
-  channelExists: (
-    channelKey: Hash,
-    from: StarknetAddress,
-    to: StarknetAddress,
-    toPublicKey: PublicKey
-  ): Hash => hash(channelKey, from, to, toPublicKey),
-  tokenKey: (channelKey: Hash, nonce: NoteNonce): Hash =>
-    hash(channelKey, nonce.slot, nonce.sequence),
-  tokenExists: (
-    channelKey: Hash,
-    to: StarknetAddress,
-    toPublicKey: PublicKey,
-    token: StarknetAddress
-  ): Hash => hash(channelKey, to, toPublicKey, token),
-  noteId: (witness: Witness, token: StarknetAddress): Hash =>
-    hash(witness.channelKey, token, witness.nonce.slot, witness.nonce.sequence),
-  nullifier: (witness: Witness, token: StarknetAddress, ownerPrivateKey: PrivateKey): Hash =>
-    hash(witness.channelKey, token, witness.nonce.slot, witness.nonce.sequence, ownerPrivateKey),
-};
