+++ sdk/src/testing/index.ts
@@ -7,4 +7,5 @@ export { PrivacyPool } from "./pool.js";
 export { MockDiscoveryProvider } from "./discovery.js";
 export { MockTokenOperationsBuilder, MockPrivateTransfersBuilder } from "./builders.js";
 export { MockPrivateTransfers } from "./transfers.js";
-export { createMockProof, createMockCallAndProof, hashes, Withdrawal } from "./helpers.js";
+export { createMockProof, createMockCallAndProof, Withdrawal } from "./helpers.js";
+export { hashes } from "../utils/hashes.js";
